# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级任务管理系统的 RESTful API 接口文档",
        "title": "任务管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.12.48:8888",
    "basePath": "/api/v1",
    "paths": {
        "/admin/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统中所有权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "获取权限列表",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dto.PermissionResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统中所有角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "获取角色列表",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dto.RoleResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "用户使用用户名/邮箱/手机号和密码登录，返回JWT令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录凭证",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，返回token",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "新用户注册，注册后需要等待管理员审核",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码（用户登录密码）",
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "description": "用户名或邮箱（用户在系统中注册的用户名或邮箱）",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT 访问令牌（用于后续API请求的身份验证）",
                    "type": "string"
                },
                "user_info": {
                    "description": "用户信息（包含用户基本信息和权限等）"
                }
            }
        },
...
        "dto.PermissionResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "权限描述",
                    "type": "string"
                },
                "id": {
                    "description": "权限ID",
                    "type": "integer"
                },
                "name": {
                    "description": "权限名称",
                    "type": "string"
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱地址（唯一，格式必须为有效邮箱）",
                    "type": "string",
                    "example": "user@example.com"
                },
                "mobile": {
                    "description": "手机号码（可选，格式为11位中国手机号）",
                    "type": "string",
                    "example": "13800138000"
                },
                "nickname": {
                    "description": "昵称（用户昵称，用于显示，可选）",
                    "type": "string",
                    "example": "John"
                },
                "password": {
                    "description": "密码（6-20个字符，需要包含字母和数字）",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6,
                    "example": "password123"
                },
                "username": {
                    "description": "用户名（唯一，字母数字组合，5-50个字符）",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "dto.RoleResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "角色描述",
                    "type": "string"
                },
                "id": {
                    "description": "角色ID",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "permissions": {
                    "description": "角色包含的权限列表（可选）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PermissionResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}